Add support for telnet TIMING-MARK option.
Send telnet TIMING-MARK option once in the command queue before other
initial option requests; enable acknowledgement mode if and only if a
response to that option returns before the ECHO option returns.  When
enabled, the option is used to generate an end-to-end acknowledgement
from the remote telnet client, used to verify which output has likely
been "seen".  Even if the option is not understood by the remote end,
the client *should* reject the option according to standard procedure
defined by the telnet protocol specifications.  If the option is
rejected, it is nearly as useful, since it still does guarantee that
the telnet client at least received the data instead of losing it over
the network.  If the telnet client is broken and does not respond to
the option at all, then acknowledgements cannot be used and data might
be lost by the network.