Use TIMING-MARK acknowledgements in output stream.
Update output stream buffering code to allow for variable-sized output
window using acknowledgements based on the standard telnet TIMING-MARK
option.  If the telnet client is broken and doesn't respond to that
option at all, fake "acknowledgements" are generated instead when a
write() accepts all output data into the kernel buffers.  Effectively,
this limits the window size of the output stream to one.  The network
could lose data if the client does not respond to the option.